Removed intermediary payloadExec step.
All the extra step was doing was verifying that there was at least one
instruction. Considering how many ways payloadExec can fail this check
seemed unnecessary. Removing the check freed three registers, one of
which was used to put a jump to SendResponseAndResetPointer in the fixed
area. The other two were added to the static buffer.